Positioned Photos / studentSuccess Text Correctly
Added foundation push and pull classes to all student success sections
of text. This will position photos above the text on the phone and
position photos to the right or left of the text in an alternating
series. Also increased the size of the fonts on both the titles and the
text.